Upgrade packages, fix dart errors

+++ noterly/lib/widgets/first_launch_dialog.dart
@@ -148,7 +148,7 @@ class FirstLaunchDialog extends Dialog {
               ),
               DotsIndicator(
                 dotsCount: pages.length,
-                position: currentPage,
+                position: currentPage.toDouble(),
                 mainAxisAlignment: MainAxisAlignment.center,
               ),
               ButtonBar(
